config-linux: Default should be ENOSYS.
On Linux the major C libraries expect that syscalls that are
blocked from running in the container runtime return ENOSYS
to allow fallbacks to be used. Returning EPERM by default is
not useful particularly for syscalls that would return EPERM
for actual access restrictions e.g. the new faccessat2.

The runtime-spec should set the standard and recommend ENOSYS
be returned just like a kernel would that doesn't support that
syscall. This allows C runtimes to fall back on other possible
implementations given the userspace policies.

Please see the upstream discussions:
https://lwn.net/Articles/738694/
- Discusses fragility of syscall filtering.
https://github.com/opencontainers/runc/issues/2151
- glibc and musl request ENOSYS return for unknown syscalls.
https://github.com/systemd/systemd/pull/16739
- Discusses systemd-nspawn breakage with faccessat2.
https://github.com/systemd/systemd/pull/16819
- General policy for systemd-nspawn to return ENOSYS.
https://github.com/seccomp/libseccomp/issues/286
- Block unknown syscalls and return ENOSYS.